moved components to view folder
