+++ src/main/java/org/cmps/tetrahedron/view/component/FilePicker.java
@@ -1,4 +1,4 @@
-package org.cmps.tetrahedron.view;
+package org.cmps.tetrahedron.view.component;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
@@ -6,13 +6,10 @@ import javafx.stage.FileChooser;
 import lombok.Getter;
 import lombok.Setter;
 import org.cmps.tetrahedron.controller.FileChooserController;
-import org.cmps.tetrahedron.controller.LocalizationController;
 import org.cmps.tetrahedron.controller.SceneController;
-import org.cmps.tetrahedron.i18n.LocalizationListener;
 
 import java.io.File;
 import java.util.ResourceBundle;
-import java.util.prefs.Preferences;
 
 /**
  * TODO: add description.
@@ -29,14 +26,9 @@ public class FilePicker {
     private Label label;
 
     private File file;
-    private static final String LAST_USED_DIRECTORY_KEY = "last_used_directory";
-    private static final Preferences prefs = Preferences.userNodeForPackage(FilePicker.class);
 
+    @Setter
     private Runnable onFileSelectedCallback;
-
-    public void setOnFileSelectedCallback(Runnable callback) {
-        this.onFileSelectedCallback = callback;
-    }
 
     @FXML
     public void onClick() {
@@ -56,5 +48,4 @@ public class FilePicker {
     public void showNotSelectedFileError() {
         label.setText(resources.getString("file-is-not-selected"));
     }
-
 }
